+++ frontend/src/features/feed/components/Highlights.tsx
@@ -1,0 +1,16 @@
+export function Highlights({ items }: { items?: string[] }) {
+  if (!items || items.length === 0) return null;
+  return (
+    <div className="mt-2 flex flex-wrap gap-1.5">
+      {items.map((t) => (
+        <span
+          key={t}
+          className="inline-flex items-center rounded-full bg-secondary px-2 py-0.5 text-[11px] leading-4 text-secondary-foreground ring-1 ring-border"
+        >
+          {t}
+        </span>
+      ))}
+    </div>
+  );
+}
+
